requirements.txt
pytest==8.3.3
pytest-cov==6.0.0
pandas==2.2.3
python-dotenv==1.0.1
colorama==0.4.6
